+++ .eslintrc.js
@@ -19,6 +19,7 @@ module.exports = {
     ],
     env: {
       jest: true,
+      'jest/globals': true,
     },
     rules: {
       'jsx-a11y/label-has-for': 'off',