+++ src/components/custom-field-input-single-choice/custom-field-input-single-choice.jsx
@@ -7,39 +7,37 @@ import React, {
   useState,
 } from 'react';
 import PropTypes from 'prop-types';
+import useFetch from '@bloodyaugust/use-fetch';
 import Select from '../select/select.jsx';
 import ListOption from '../list-option/list-option.jsx';
+import mockConstants from '../../mocks/mock-constants.js';
+
+const { API_ROOT } = mockConstants;
+const defaultValue = ['-1'];
 
 const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleChoice(props, ref) {
+  const [choices, setChoices] = useState([]);
+  const listOptionRefs = choices.map(() => createRef());
   const [value, setValue] = useState(props.value);
   const backupRef = useRef();
   const selfRef = ref || backupRef;
-
-  const listOptionRefs = props.choices.map(() => createRef());
+  const { execute } = useFetch();
 
   function selectOnChangeHandler(event) {
     if (event.target.value) {
-      setValue(props.choices.find(choice => choice.id === event.target.value.id));
+      setValue([event.target.value.id]);
     } else {
-      setValue(undefined);
+      setValue(defaultValue);
     }
   }
 
   const listOptions = () => {
-    const choices = {};
-    props.choices.forEach((item, index) => { choices[item.id] = { ...item, index }; });
-
-    return props.choices
-      .map(item => choices[item.id])
-      .map(item => (
+    return choices
+      .map((item, index) => (
         <ListOption
           key={item.id}
-          ref={listOptionRefs[item.index]}
-          selected={value && item.id === value.id}
-          value={{
-            id: item.id,
-            label: item.label,
-          }}
+          ref={listOptionRefs[index]}
+          value={item}
         >
           {item.label}
         </ListOption>
@@ -48,73 +46,100 @@ const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleC
 
   useImperativeHandle(selfRef, () => ({
     get dirty() {
-      const providedValue = props.value ? props.value.id : undefined;
+      const providedValue = props.value[0] !== defaultValue[0] ? props.value[0] : undefined;
       return providedValue !== this.value[0];
     },
     id: props.id,
     name: props.name,
     get value() {
-      return value ? [value.id] : [];
+      return value[0] === defaultValue[0] ? [] : value;
     },
   }));
 
   useEffect(() => {
-    setValue(props.value);
-  }, [props.value]);
+    if (props.value.length) {
+      setValue(props.value);
+    } else {
+      setValue(defaultValue);
+    }
+  }, (props.value.length ? [...props.value] : [...defaultValue]));
 
   useEffect(() => {
-    if (props.value === value) return;
-
     props.onChange({ target: selfRef.current });
   }, [value]);
 
-  return (
-    <Select
-      className={props.className}
-      displayValueEvaluator={selectValue => selectValue.label}
-      errorText={props.errorText}
-      id={props.id}
-      label={props.label}
-      listOptionRefs={listOptionRefs}
-      name={props.name}
-      onChange={selectOnChangeHandler}
-      placeholder={props.placeholder}
-      readOnly={props.readOnly}
-      required={props.required}
-      value={value}
-    >
-      { listOptions() }
-    </Select>
-  );
-});
+  useEffect(() => {
+    const getChoices = async () => {
+      await execute(`${API_ROOT}/custom_field_choices?for_custom_fields=${props.customFieldID}`)
+        .then(({ json, mounted }) => {
+          if (mounted) {
+            const mungedChoices = json.results.map((result) => {
+              return {
+                ...json[result.key][result.id],
+                id: result.id,
+              };
+            });
 
-const ChoiceType = PropTypes.shape({
-  id: PropTypes.number.isRequired,
-  label: PropTypes.string.isRequired,
+            setChoices(mungedChoices);
+          }
+        })
+        .catch((error) => {
+          if (error.error && error.error.type !== 'aborted') {
+            throw error;
+          }
+        });
+    };
+
+    getChoices();
+  }, []);
+
+  return (
+    <React.Fragment>
+      <Select
+        className={props.className}
+        displayValueEvaluator={selectValue => selectValue.label}
+        errorText={props.errorText}
+        id={props.id}
+        label={props.label}
+        listOptionRefs={listOptionRefs}
+        name={props.name}
+        onChange={selectOnChangeHandler}
+        placeholder={props.placeholder}
+        readOnly={props.readOnly}
+        required={props.required}
+        value={value[0] !== -1 ? choices.find(choice => choice.id === value[0]) : undefined}
+      >
+        { listOptions() }
+      </Select>
+      {choices.length === 0 &&
+        <span>Loading...</span>
+      }
+    </React.Fragment>
+  );
 });
 
 CustomFieldInputSingleChoice.propTypes = {
   id: PropTypes.string.isRequired,
-  choices: PropTypes.arrayOf(ChoiceType),
   className: PropTypes.string,
+  /** Needs to be the Custom Field ID from the database for self-loading of choices */
+  customFieldID: PropTypes.string.isRequired,
   label: PropTypes.string.isRequired,
   name: PropTypes.string.isRequired,
   onChange: PropTypes.func,
   placeholder: PropTypes.string,
   readOnly: PropTypes.bool,
   required: PropTypes.bool,
-  value: ChoiceType,
+  value: PropTypes.arrayOf(PropTypes.string),
   errorText: PropTypes.string,
 };
 
 CustomFieldInputSingleChoice.defaultProps = {
-  choices: [],
   className: undefined,
   onChange: () => {},
   placeholder: undefined,
   readOnly: false,
   required: false,
-  value: undefined,
+  value: defaultValue,
   errorText: undefined,
 };