+++ src/mocks/jest-server.js
@@ -1,0 +1,5 @@
+import { setupServer } from 'msw/node'; // eslint-disable-line import/extensions
+
+const jestServer = setupServer();
+
+export default jestServer;