+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
+  - Major: Add self-loading of choices to `CustomFieldSingleChoice`
+  - Major: Remove `choices` from props API of `CustomFieldSingleChoice`
+  - Minor: Add `customFieldID` prop to `CustomFieldSingleChoice`
+  - Major: Change `value` prop of `CustomFieldSingleChoice` to accept an array of IDs
+  - Minor: Bugfix for `ListOption` to change class based on selected state, not selected prop only
 </details>
 
 ## v0.60.0